Add more structure to the benchmarking method
